Pull build_canister_methods down to query/update
